LayoutTests/fast/css/variables/color-hex-test-expected.txt

CONSOLE MESSAGE: line 27: TypeError: 'undefined' is not an object (evaluating 'declaration.getVariableValue')
FAILED
